DNS Hold-on proxy
A shield against DNS poisoning that happens in censored networks. 
The first packet that is received by stub/proxy is not taken as default response if it doesn't fit certain rtt and tll criteria. We wait till the end of a hold-on period and second packet that is received is chosen as the valid packet.